refactor for better structure visibility
