+++ mobile/src/main/java/com/example/samplewearmobileapp/models/FixedSizeList.kt
@@ -1,4 +1,4 @@
-package com.example.samplewearmobileapp.model
+package com.example.samplewearmobileapp.models
 
 import java.util.*
