+++ mobile/src/main/java/com/example/samplewearmobileapp/models/PlotArrays.kt
@@ -1,4 +1,4 @@
-package com.example.samplewearmobileapp.model
+package com.example.samplewearmobileapp.models
 
 /**
  * Class to hold arrays for plotting.
