+++ wear/src/main/java/com/example/samplewearmobileapp/ConnectionManager.kt
@@ -4,6 +4,10 @@ import android.content.Context
 import android.os.Handler
 import android.os.Looper
 import android.util.Log
+import com.example.samplewearmobileapp.trackers.Listener
+import com.example.samplewearmobileapp.trackers.ppggreen.PpgGreenListener
+import com.example.samplewearmobileapp.trackers.ppgir.PpgIrListener
+import com.example.samplewearmobileapp.trackers.ppgred.PpgRedListener
 import com.samsung.android.service.health.tracking.ConnectionListener
 import com.samsung.android.service.health.tracking.HealthTrackerException
 import com.samsung.android.service.health.tracking.HealthTrackingService
