+++ wear/src/main/java/com/example/samplewearmobileapp/TrackerDataNotifier.kt
@@ -1,5 +1,8 @@
 package com.example.samplewearmobileapp
 
+import com.example.samplewearmobileapp.trackers.ppggreen.PpgGreenData
+import com.example.samplewearmobileapp.trackers.ppgir.PpgIrData
+import com.example.samplewearmobileapp.trackers.ppgred.PpgRedData
 import java.util.function.Consumer
 
 class TrackerDataNotifier {
