+++ wear/src/main/java/com/example/samplewearmobileapp/TrackerDataObserver.kt
@@ -1,5 +1,9 @@
 package com.example.samplewearmobileapp
 
+import com.example.samplewearmobileapp.trackers.ppggreen.PpgGreenData
+import com.example.samplewearmobileapp.trackers.ppgir.PpgIrData
+import com.example.samplewearmobileapp.trackers.ppgred.PpgRedData
+
 interface TrackerDataObserver {
 //    fun onHeartRateTrackerDataChanged(hrData: HeartRateData)
     fun onPpgGreenTrackerDataChanged(ppgGreenData: PpgGreenData)
