+++ wear/src/main/java/com/example/samplewearmobileapp/trackers/Listener.kt
@@ -1,4 +1,4 @@
-package com.example.samplewearmobileapp
+package com.example.samplewearmobileapp.trackers
 
 import android.os.Handler
 import android.util.Log
