+++ wear/src/main/java/com/example/samplewearmobileapp/trackers/ppgred/PpgRedListener.kt
@@ -1,25 +1,22 @@
-package com.example.samplewearmobileapp
+package com.example.samplewearmobileapp.trackers.ppgred
 
 import android.util.Log
+import com.example.samplewearmobileapp.R
+import com.example.samplewearmobileapp.TrackerDataNotifier
+import com.example.samplewearmobileapp.trackers.Listener
 import com.samsung.android.service.health.tracking.HealthTracker
 import com.samsung.android.service.health.tracking.data.DataPoint
 import com.samsung.android.service.health.tracking.data.ValueKey
 
-class PpgGreenListener internal constructor() : Listener() {
-    private val tag = "PpgGreenListener"
+class PpgRedListener internal constructor() : Listener() {
+    private val tag = "PpgRedListener"
 
     init {
         val trackerEventListener: HealthTracker.TrackerEventListener = object :
             HealthTracker.TrackerEventListener {
             override fun onDataReceived(list: List<DataPoint>) {
-                if (list.isNotEmpty()) {
-                    Log.i(tag, "Ppg Data List Size : " + list.size)
-                    for (dataPoint in list) {
-                        readValuesFromDataPoint(dataPoint)
-                    }
-                } else {
-                    Log.i(tag, "onDataReceived List is empty")
-                    readZeroValue()
+                for (data in list) {
+                    readValuesFromDataPoint(data)
                 }
             }
 
@@ -45,25 +42,14 @@ class PpgGreenListener internal constructor() : Listener() {
         Log.i(tag, "Timestamp : " + dataPoint.timestamp)
         Log.i(
             tag,
-            "Ppg Green Value : " + dataPoint.getValue(ValueKey.PpgGreenSet.PPG_GREEN)
+            "Ppg InfraRed Value : " + dataPoint.getValue(ValueKey.PpgRedSet.PPG_RED)
         )
 
-        val ppgGreenData = PpgGreenData()
-        ppgGreenData.status = PpgGreenStatus.PPG_GREEN_STATUS_GOOD.code
-        ppgGreenData.ppgValue = dataPoint.getValue(ValueKey.PpgGreenSet.PPG_GREEN)
-        ppgGreenData.timestamp = dataPoint.timestamp
+        val ppgRedData = PpgRedData()
+        ppgRedData.ppgValue = dataPoint.getValue(ValueKey.PpgRedSet.PPG_RED)
+        ppgRedData.timestamp = dataPoint.timestamp
 
-        TrackerDataNotifier.instance?.notifyPpgGreenTrackerObservers(ppgGreenData)
+        TrackerDataNotifier.instance?.notifyPpgRedTrackerObservers(ppgRedData)
         Log.d(tag, dataPoint.toString())
-    }
-
-    fun readZeroValue() {
-        val zeroPpgGreenData = PpgGreenData(
-            PpgGreenStatus.PPG_GREEN_STATUS_NONE.code,
-            0,
-            0
-        )
-        TrackerDataNotifier.instance?.notifyPpgGreenTrackerObservers(zeroPpgGreenData)
-        Log.d(tag, "Zero PPG Value notified")
     }
 }
